+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   * Implemented MPAddAreaTHIDCommand test.
   * Renamed MPTH2FileEditStateSingleLineEdit to MPTH2FileEditStateEditSingleLine
   * flutter upgrade to 3.35.5
+  * Created MPCommandFactory.addXTherionInsertImageConfig().
   * Added tests:
     * add area THID that creates area and line THIDs;
     * add area;